feat: square components also accepted in vgg-repeat
