+++ src/components/Core/Map.vue
@@ -2,6 +2,7 @@
 import DataReceiver from '../../mixins/Data/DataReceiver.js'
 import CoordinateTreeUser from '../../mixins/CoordinateTreeUser.js'
 import ScaleReceiver from '../../mixins/Scales/ScaleReceiver.js'
+import isSquareComponent from './utils/isSquareComponent.js'
 
 import {
   calculateGridLayout,
@@ -84,16 +85,7 @@ export default {
       }
     },
 
-    isSquareComponent (component) {
-      let props = component.componentOptions.Ctor.options.props
-      let squareProps = ['x1', 'x2', 'y1', 'y2']
-      for (let prop of squareProps) {
-        if (!props[prop]) {
-          return false
-        }
-      }
-      return true
-    }
+    isSquareComponent
   },
 
   provide () {
